requirements.txt
streamlit
tensorflow
tensorflow_hub
Pillow